pyperc library deep clean
I wrote this, originally, as a quick hack in php.  Later it was
converted to Python, but a fairly direct conversion, and not Pythonic at
all.

Now, the SortedCollection provides a minimalist ram database function,
the high-level PyPerc class is separated from the MegaCLI parsers, and
the CLI runner itself is also split to assist with mocks.

The events are loaded directly from the card on load, and I avoid
Clearing the log, and instead read with -GetLatest N, and try to make N
large enough to avoid missing any data (but also detect if we did)